Overwrote PrologGrammar.java with the one from jkbotev's directory.
Corrected a casting error on line 170, changing (Double) to (Long), which is being returned by maybeName.second().

Getting runtime existence errors such as

/(notation, 3) does not exist.

+++ src/imp/attributes/PrologGrammar.java
@@ -167,7 +167,7 @@ public class PrologGrammar {
         if (givenExpansion.isEmpty()) {
             name = new StructureTerm(
                     SymbolTerm.makeSymbol((String) maybeName.first(), 1),
-                    new Term[] { new DoubleTerm((Double) maybeName.second())});
+                    new Term[] { new DoubleTerm((Long)maybeName.second())});
             return new StructureTerm(SymbolTerm.makeSymbol("rule",4),
                 new Term[] {
                   name,
